update llama error

+++ code/lavis/models/blip2_models/mini_gpt4.py
@@ -92,13 +92,13 @@ class MiniGPT4(Blip2Base):
                 torch_dtype=torch.float16,
                 load_in_8bit=True,
                 device_map={'': device_8bit}
-            )
+            ).bfloat16()
         else:
             self.llama_model = LlamaForCausalLM.from_pretrained(
                 llama_model,
                 torch_dtype=torch.float16,
                 low_cpu_mem_usage=True
-            )
+            ).bfloat16()
 
         for name, param in self.llama_model.named_parameters():
             param.requires_grad = False
